requirements.txt
qq-botpy
psutil
requests
lxml
openai